chore: add POST /v1/settings.bulk
Extract saveSettingsBulk into app/lib/server/functions/ and call
it from both REST and DDP. The endpoint is auth-gated, enforces
2FA (disableRememberMe), and reuses the per-setting permission
chain (edit-privileged-setting OR manage-selected-settings + per
-id permission) the DDP method already enforced.
